add all hardware table
