+++ src/translations/en.json
@@ -4211,6 +4211,7 @@
             "failed_to_get": "Failed to get available hardware",
             "title": "All hardware",
             "search": "Search hardware",
+            "name": "Name",
             "subsystem": "Subsystem",
             "device_path": "Device path",
             "id": "ID",
